Fixed compiler warnings.

+++ src/core/xt-aggregate.cpp
@@ -217,7 +217,7 @@ XtFault XtAggregate::Stop() {
   XtError result = 0;
   XtiCas(&running, 0, 1);
   while(XtiCas(&insideCallbackCount, 0, 0) != 0);
-  if(masterIndex == -1 || masterIndex >= streams.size())
+  if(masterIndex == -1 || static_cast<size_t>(masterIndex) >= streams.size())
     return 0;
   if((error = XtStreamStop(streams[masterIndex].get())) != 0)
     result = error;
